Extracted loading of cloning facilities into a DataLoader object, to reduce the complexity of CloningService
