+++ src/main/java/com/projectswg/holocore/resources/support/data/server_info/loader/ServerData.kt
@@ -107,6 +107,7 @@ object ServerData {
 	val rawResources		by SoftDataLoaderDelegate(::RawResourceLoader)
 	val staticCities		by SoftDataLoaderDelegate(::StaticCityLoader)
 	val explorationBadges	by SoftDataLoaderDelegate(::ExplorationBadgeLoader)
+	val cloningFacilities	by SoftDataLoaderDelegate(::CloningFacilityLoader)
 
 	private class WeakDataLoaderDelegate<T: DataLoader>(loaderCreator: () -> T): DataLoaderDelegate<T>(::WeakReference, loaderCreator)
 	private class SoftDataLoaderDelegate<T: DataLoader>(loaderCreator: () -> T): DataLoaderDelegate<T>(::SoftReference, loaderCreator)
